refactor(Result) : 플레이어들의 결과를 맵을 사용하여 관리하도록 변경

+++ src/main/java/blackjack/domain/BettingAmount.java
@@ -20,6 +20,6 @@ public class BettingAmount {
     }
 
     public Profit calculateProfit(ResultStatus resultStatus) {
-        return new Profit(resultStatus.calculateProfit(value));
+        return new Profit((int) resultStatus.calculateProfit(value));
     }
 }
